Moved LogLevel Outside Internal folder
